Primeros archivos

+++ com.Agenda/src/java/com/Entidades/Agenda.java
@@ -47,7 +47,7 @@ public class Agenda implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "ID_AGENDA")
-    private Long idAgenda;
+    private Integer idAgenda;
     @Size(max = 15)
     @Column(name = "TIPO")
     private String tipo;
@@ -78,15 +78,15 @@ public class Agenda implements Serializable {
     public Agenda() {
     }
 
-    public Agenda(Long idAgenda) {
+    public Agenda(Integer idAgenda) {
         this.idAgenda = idAgenda;
     }
 
-    public Long getIdAgenda() {
+    public Integer getIdAgenda() {
         return idAgenda;
     }
 
-    public void setIdAgenda(Long idAgenda) {
+    public void setIdAgenda(Integer idAgenda) {
         this.idAgenda = idAgenda;
     }
